* dwarf2out.c (lookup_filename): Fix printf format warning.         * system.h (fread_unlocked, fwrite_unlocked): Undef.
        * fixinc/Makefile.in (FL_LIST): Add $($@-warn) hook.
        (fixincl.o-warn, gnu-regex.o-warn): New.
        * fixinc/fixfixes.c (FIX_PROC_HEAD): Mark parameters unused.
        * fixinc/fixtests.c (TEST_FOR_FIX_PROC_HEAD): Likewise.
        * fixinc/fixincl.c (process): Fix printf format warning.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@61584 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/dwarf2out.c
@@ -12354,7 +12354,7 @@ lookup_filename (file_name)
 
   if (DWARF2_ASM_LINE_DEBUG_INFO)
     {
-      fprintf (asm_out_file, "\t.file %u ", i);
+      fprintf (asm_out_file, "\t.file %lu ", (unsigned long) i);
       output_quoted_string (asm_out_file, file_name);
       fputc ('\n', asm_out_file);
     }
